modified DAG for demographics data to ingest into tables in "development" schema
